testing(graphql, rest): add preflight specific tests (#46)
