+++ Test/Integration/Response/GraphQlResponseTest.php
@@ -3,14 +3,14 @@
  * Copyright © Graycore, LLC. All rights reserved.
  * See LICENSE.md for details.
  */
-namespace Graycore\Cors\Test\Integration;
+namespace Graycore\Cors\Test\Integration\Response;
 
 use Magento\Framework\App\Response\Http;
 use Magento\TestFramework\TestCase\AbstractController as ControllerTestCase;
 use Zend\Http\Headers;
 
 /**
- * Tests that the responses to REST API requests
+ * Tests that the responses to GraphQl API requests
  * properly respond with the CORS headers in the
  * default configuration
  * @author    Graycore <damien@graycore.io>
@@ -18,58 +18,39 @@ use Zend\Http\Headers;
  * @license   MIT https://github.com/graycoreio/magento2-cors/license
  * @link      https://github.com/graycoreio/magento2-cors
  */
-class WebApiResponseTest extends ControllerTestCase
+class GraphQlResponseTest extends ControllerTestCase
 {
-    const ENDPOINT = '/rest/default/V1/directory/currency';
-
-    public function getResponse()
-    {
-        if (!$this->_response) {
-            $this->_response = $this->_objectManager->get(\Magento\Framework\Webapi\Rest\Response::class);
-        }
-        return $this->_response;
-    }
-
-    private function dispatchToRestApi()
-    {
-        ob_start();
-        $this->dispatch(self::ENDPOINT);
-        ob_end_clean();
-    }
-
-    private function dispatchToRestApiWithOrigin(string $origin)
+    private function dispatchToGraphQlApiWithOrigin(string $origin)
     {
         $headers = new Headers();
         $headers->addHeaderLine('Origin: ' . $origin);
         $headers->addHeaderLine('Content-Type: application/json');
-        $this->getRequest()->setMethod('GET')->setHeaders($headers);
-        ob_start();
-        $this->dispatch(self::ENDPOINT);
-        ob_end_clean();
+        $this->getRequest()->setMethod('POST')->setHeaders($headers)
+            ->setContent('{"query": "{products(search:\"Pierce\"){total_count}}"}');
+        $this->dispatch('/graphql');
     }
 
     /**
-     * @magentoConfigFixture default/web/api_rest/cors_allowed_origins https://www.example.com
+     * @magentoConfigFixture default/web/graphql/cors_allowed_origins https://www.example.com
      */
-    public function testTheRestApiResponseContainsCrossOriginHeaders()
+    public function testTheGraphQlResponseContainsCrossOriginHeaders()
     {
-        $this->dispatchToRestApiWithOrigin("https://www.example.com");
+        $this->dispatchToGraphQlApiWithOrigin("https://www.example.com");
 
         /** @var Http $response */
         $response = $this->getResponse();
-
         $this->assertNotFalse($response->getHeader('Access-Control-Allow-Origin'));
         $this->assertNotFalse($response->getHeader('Access-Control-Max-Age'));
     }
 
     /**
-     * @magentoConfigFixture default/web/api_rest/cors_allowed_origins https://www.example.com
+     * @magentoConfigFixture default/web/graphql/cors_allowed_origins https://www.example.com
      */
     public function testItdoesNotAddAnyCrossOriginHeadersToATypicalRequest()
     {
         $headers = new Headers();
         $headers->addHeaderLine('Origin: https://www.example.com');
-        $this->dispatchToRestApi();
+        $this->dispatch('/');
 
         /** @var Http $response */
         $response = $this->getResponse();
@@ -78,7 +59,7 @@ class WebApiResponseTest extends ControllerTestCase
 
     public function testItDoesNotAddAnyCrossOriginHeadersOutOfTheBox()
     {
-        $this->dispatchToRestApiWithOrigin("https://www.example.com");
+        $this->dispatchToGraphQlApiWithOrigin("https://www.example.com");
 
         /** @var Http $response */
         $response = $this->getResponse();
@@ -86,14 +67,14 @@ class WebApiResponseTest extends ControllerTestCase
         $this->assertFalse($response->getHeader('Access-Control-Max-Age'));
     }
 
-    public function testTheRestApiWillRespondToAOptionsRequestWithA200Response()
+    public function testTheGraphQlApiWillRespondToAOptionsRequestWithA200Response()
     {
         $headers = new Headers();
         $headers->addHeaderLine('Origin: https://www.example.com');
         $headers->addHeaderLine('Content-Type: application/json');
 
         $this->getRequest()->setMethod('OPTIONS')->setHeaders($headers);
-        $this->dispatchToRestApi();
+        $this->dispatch('/graphql');
 
         /** @var Http $response */
         $response = $this->getResponse();
@@ -101,17 +82,16 @@ class WebApiResponseTest extends ControllerTestCase
     }
 
     /**
-     * @group fit
-     * @magentoConfigFixture default/web/api_rest/cors_allowed_origins https://www.example.com
+     * @magentoConfigFixture default/web/graphql/cors_allowed_origins https://www.example.com
      */
-    public function testTheRestApiWillRespondToAOptionsRequestWithCorsHeadersOnTheResponse()
+    public function testTheGraphQlApiWillRespondToAOptionsRequestWithCorsHeadersOnTheResponse()
     {
         $headers = new Headers();
         $headers->addHeaderLine('Origin: https://www.example.com');
         $headers->addHeaderLine('Content-Type: application/json');
 
-        $this->getRequest()->setMethod('POST')->setHeaders($headers);
-        $this->dispatch(self::ENDPOINT);
+        $this->getRequest()->setMethod('OPTIONS')->setHeaders($headers);
+        $this->dispatch('/graphql');
 
         /** @var Http $response */
         $response = $this->getResponse();
